Moved Delete Snackbar to Grid
